requirements.txt
robyn==0.66.2
pydantic==2.11.3
openai==1.76.2
email-validator==2.2.0
python-dotenv==1.1.0
openai-agents==0.0.14